+++ web/src/components/design-system/Radio/Radio.stories.tsx
@@ -1,5 +1,5 @@
 import { fn } from "storybook/test";
-import preview from "../../../.storybook/preview";
+import preview from "../../../../.storybook/preview";
 import { Radio } from "./Radio";
 
 const meta = preview.meta({
